requirements.txt
websocket-client==0.37.0
